test version beta1.2

+++ app/src/main/java/com/crowdos/ui/notifications/YourFollower.java
@@ -6,6 +6,6 @@ public class YourFollower {
     public double longitude;
     public double latitude;
     public long startTime;
-    public int eventId;
+    public long eventId;
     boolean eventType;
 }
